Make BitVec be generic over a storage implementing the Storage trait.
Before this commit, BitVec<B> was backed by a Vec<B>. However, this
commit introduces the Storage trait with the associated type B: BitBlock
and some vec-like methods. BitVec<S> is now generic over the storage S
instead.

The storage trait is implemented for Vec, and the optional features
arrayvec, smallvec, and allocator-api2 implements it for ArrayVec<B, N>,
SmallVec<[B; N]>, and allocator_api2::vec::Vec<B, A> respectively.

This change is obviously not backwards compatable as it changes the
signiture of the BitVec struct.